django/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
backoff==2.2.1
blinker==1.6.2
bottle==0.12.25
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.5
Deprecated==1.2.14
Django==4.2.2
exceptiongroup==1.1.2
fastapi==0.100.0
Flask==2.3.2
googleapis-common-protos==1.59.1
grpcio==1.56.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hupper==1.12
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-aws-lambda==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-django==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-tortoiseorm==0.39b0
opentelemetry-instrumentation-urllib==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
PasteDeploy==3.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
protobuf==4.23.3
pydantic==2.0.3
pydantic_core==2.3.0
pyramid==2.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tornado==6.3.2
translationstring==1.4
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
venusian==3.0.0
watchfiles==0.19.0
WebOb==1.8.7
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
zope.deprecation==5.0
zope.interface==6.0
